requirements.txt
alembic==1.7.6
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
flask_postgres==0.2.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
psycopg2==2.9.3
python-dotenv==0.19.2
SQLAlchemy==1.4.31
Werkzeug==2.0.3
wtf==0.1
WTForms==3.0.1
